reestructured the tree-project
