--- Bow OpenAPI/Bow OpenAPI/Utils/FileSystemError+Extensions.swift
@@ -1,9 +1,0 @@
-//  Copyright © 2019 The Bow Authors.
-
-import Foundation
-
-extension FileSystemError {
-    static func toAPIClientError(_ error: FileSystemError) -> APIClientError {
-        APIClientError(operation: "FileSystem", error: error)
-    }
-}
